fix: 修复防盗链处理功能部分问题 - 提取 shouldAddNoReferrer 到 image-utils.ts 消除三处重复实现 - 修复 rehype-figure 匹配域名时错误清空 alt 文本导致 figcaption 丢失 - 移除 Layout 客户端脚本中清空 src 重载图片的逻辑以避免闪烁 - 跳过非 http URL 避免对本地图片做无意义的 URL 解析
