+++ src/components/common/CoverImage.astro
@@ -2,12 +2,13 @@
 import { Picture } from "astro:assets";
 import * as path from "node:path";
 import type { ImageMetadata } from "astro";
-import { coverImageConfig, siteConfig } from "@/config";
+import { coverImageConfig } from "@/config";
 import type { ImageFormat, ResponsiveImageLayout } from "@/types/config";
 import {
 	getFallbackFormat,
 	getImageFormats,
 	getImageQuality,
+	shouldAddNoReferrer,
 } from "@/utils/image-utils";
 import { url } from "@/utils/url-utils";
 
@@ -50,28 +51,6 @@ const isLocal = !(
 	src.startsWith("data:")
 );
 const isPublic = src.startsWith("/");
-
-// 检查是否需要为图片添加 referrerpolicy
-function shouldAddNoReferrer(urlStr: string): boolean {
-	const domains = siteConfig.imageOptimization?.noReferrerDomains || [];
-	if (domains.length === 0) return false;
-
-	try {
-		const urlObj = new URL(
-			urlStr,
-			urlStr.startsWith("http") ? undefined : "http://localhost",
-		);
-		const hostname = urlObj.hostname;
-
-		return domains.some((pattern) => {
-			const regexPattern = pattern.replace(/\./g, "\\.").replace(/\*/g, ".*");
-			const regex = new RegExp(`^${regexPattern}$`);
-			return regex.test(hostname);
-		});
-	} catch {
-		return false;
-	}
-}
 
 // 动态导入本地图片
 let img: ImageMetadata | null = null;
